+++ hijiki/broker/hijiki_broker.py
@@ -1,20 +1,10 @@
-import os
-
 from celery import Celery
-from kombu import Exchange, Queue
 
 from hijiki.broker.broker_data import get_broker_url, init_os_environ
 
 
 class HijikiBroker:
     celery_broker = None
-
-    def bind_queues(self, queues_names):
-        for name in queues_names:
-            dlx = Exchange(f'dlq_{name}_exchange', type='topic')
-
-            dead_letter_queue = Queue(f'dlq_{name}', dlx, routing_key='*', queue_arguments={'x-queue-type': 'quorum'})
-            dead_letter_queue.bind(self.celery_broker.broker_connection()).declare()
 
     def __init__(self, app_name, host, username, password, port, cluster_hosts):
         init_os_environ(host, username, password, port, cluster_hosts)
