+++ tests/test_publisher_consumer_test.py
@@ -1,39 +1,8 @@
-import threading
 import time
 import unittest
 
-from hijiki.broker.hijiki_rabbit import HijikiQueueExchange, HijikiRabbit
 from hijiki.publisher.Publisher import Publisher
-
-result_event_list = []
-
-
-class Runner():
-    qs = [HijikiQueueExchange('teste1', 'teste1_event'), HijikiQueueExchange('teste2', 'teste2_event')]
-    gr = HijikiRabbit().with_queues_exchange(qs) \
-        .with_username("user") \
-        .with_password("pwd") \
-        .with_host("localhost") \
-        .with_port(5672) \
-        .build()
-
-    threads = []
-
-    @gr.task(queue_name="teste1")
-    def internal_consumer(data):
-        print(f"consumiu o valor:{data}")
-        result_event_list.append('recebeu evento')
-
-    def run(self):
-        t = threading.Thread(target=self.gr.run)
-        self.threads.append(t)
-        t.start()
-
-    def stop(self):
-        self.gr.terminate()
-
-    def __del__(self):
-        super()
+from tests.runner import Runner
 
 
 class TestPublisherConsumer(unittest.TestCase):
@@ -56,4 +25,4 @@ class TestPublisherConsumer(unittest.TestCase):
         time.sleep(5)
         self.pub.publish_message('teste1_event', '{"value": "Esta é a mensagem"}')
         time.sleep(1)
-        self.assertEqual(len(result_event_list), 1)
+        self.assertEqual(len(self.runner.get_results()), 1)
